remaining tablet manager test cases in GO
