+++ go/test/endtoend/tabletmanager/commands_test.go
